Issue #21 - Refactored the code to use specific FS blocksize
